+++ api/tests/integration/infrastructure/knex-database-performance-monitoring_test.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../db/knex-database-connection.js';
 import { config } from '../../../src/shared/config.js';
-import { asyncLocalStorage } from '../../../src/shared/infrastructure/monitoring-tools.js';
+import { executionContextManager } from '../../../src/shared/infrastructure/execution-context-manager.js';
 import { expect, sinon } from '../../test-helper.js';
 const selectQuery = knex.raw('SELECT 1 as value');
 
@@ -12,9 +12,9 @@ describe('Integration | Infrastructure | knex-database-performance-monitoring', 
 
     it('should store query count in context, but not the total time spent', async function () {
       // when
-      const store = await asyncLocalStorage.run({}, async () => {
+      const store = await executionContextManager.run({}, async () => {
         await selectQuery;
-        return asyncLocalStorage.getStore();
+        return executionContextManager.getStore();
       });
 
       // then
@@ -30,9 +30,9 @@ describe('Integration | Infrastructure | knex-database-performance-monitoring', 
 
     it('should store query count and total time spent in context', async function () {
       // when
-      const store = await asyncLocalStorage.run({}, async () => {
+      const store = await executionContextManager.run({}, async () => {
         await selectQuery;
-        return asyncLocalStorage.getStore();
+        return executionContextManager.getStore();
       });
 
       // then
